backend/foodgram/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf_extra_fields==3.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
Pillow==9.2.0
PyJWT==2.5.0
python-dotenv==0.20.0
pytz==2020.1
requests==2.28.1
sqlparse==0.3.1
